@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-carousel',
   standalone: true,
+  selector: 'app-carousel',
   imports: [],
   templateUrl: './carousel.component.html',
   styleUrl: './carousel.component.css',
