@@ -1,9 +1,10 @@
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-drawer',
   standalone: true,
-  imports: [],
+  selector: 'app-drawer',
+  imports: [CommonModule],
   templateUrl: './drawer.component.html',
   styleUrl: './drawer.component.css',
 })
